продолжаю работу с апи
