+++ index.html
@@ -42,7 +42,7 @@
   </div>
 </body>
 
-<script type="module" src="./js/main.js">
+<script type="module" src="./js/main.js" defer>
 </script>
 
 </html>
