+++ js/main.js
@@ -1,10 +1,8 @@
 "use strict";
 
 
-import {askDataServ} from "./api.js";
-import { userCommentInput, commentButton, arrOfComments } from "./variables.js";
-import { letClearForm, letDisabledButton } from "./changeElement.js";
-import { sendComment } from "./push.js";
+
+import { arrOfComments } from "./variables.js";
 import { renderApp } from "./render.js";
 
 //***************************************************************************************************************
@@ -18,30 +16,6 @@ import { renderApp } from "./render.js";
 //askDataServ();      //Заправшиваем данные с сервера и рендерим блок с комментами
 renderApp(arrOfComments);
 
-//инициализируем подписку на события клика и ввода в поля имени и комментария
-const userNameInput = document.getElementById('inputForName');          //поле ввода имени
-userNameInput.addEventListener('input', () => {
-    letDisabledButton(userNameInput.value)
-});
-userCommentInput.addEventListener('input', () => {
-    letDisabledButton(userCommentInput.value);
-});
-userNameInput.addEventListener('click', () => {
-    letClearForm(userNameInput);
-});
-userCommentInput.addEventListener('click', () => {
-    letClearForm(userCommentInput);
-});
 
-
-//инициализируем подписку на событие по клику по кнопке отправки коммента:
-commentButton.addEventListener('click', sendComment);
-
-//обработчик по клавише enter:
-document.addEventListener('keyup', (e) => {
-    if (e.code === 'Enter') {
-        sendComment();
-    }
-})
 
 console.log("It works!");
